+++ src/+mic/+ui/+device/GetSetLogical.m
@@ -207,7 +207,9 @@ classdef GetSetLogical <    mic.interface.ui.device.GetSetLogical & ...
                 this.uitDevice.build(this.hPanel, dLeft, dTop, this.dWidthBtn, this.dHeight);            
                 dLeft = dLeft + this.dWidthBtn;
                 
-                this.uitDevice.disable(); % re-enabled in setDevice()
+                if ~this.lDeviceIsSet
+                    this.uitDevice.disable(); % re-enabled in setDevice()
+                end
             end
             
             % Init button